Certificate utils test
