+++ ydb/core/client/server/grpc_server.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <ydb/core/grpc_services/auth_processor/dynamic_node_auth_processor.h>
+#include <ydb/core/security/certificate_check/dynamic_node_auth_processor.h>
 
 #include <ydb/core/protos/grpc.grpc.pb.h>
