--- ydb/core/driver_lib/run/cert_auth_props.cpp
@@ -1,30 +1,0 @@
-#include "cert_auth_props.h"
-
-namespace NKikimr {
-
-TDynamicNodeAuthorizationParams GetDynamicNodeAuthorizationParams(const NKikimrConfig::TClientCertificateAuthorization &clientSertificateAuth) {
-    TDynamicNodeAuthorizationParams certAuthConf;
-    if (!clientSertificateAuth.HasDynamicNodeAuthorization()) {
-        return certAuthConf;
-    }
-
-    const auto& dynNodeAuth = clientSertificateAuth.GetDynamicNodeAuthorization();
-    TDynamicNodeAuthorizationParams::TDistinguishedName distinguishedName;
-    for (const auto& term: dynNodeAuth.GetSubjectTerms()) {
-        auto name = TDynamicNodeAuthorizationParams::TRelativeDistinguishedName(term.GetShortName());
-        for (const auto& value: term.GetValues()) {
-            name.AddValue(value);
-        }
-        for (const auto& suffix: term.GetSuffixes()) {
-            name.AddSuffix(suffix);
-        }
-        distinguishedName.AddRelativeDistinguishedName(std::move(name));
-    }
-    if (!distinguishedName.RelativeDistinguishedNames.empty()) {
-        certAuthConf.AddCertSubjectDescription(distinguishedName);
-    }
-    certAuthConf.CanCheckNodeByAttributeCN = dynNodeAuth.GetCanCheckNodeHostByCN();
-    return certAuthConf;
-}
-
-}  //namespace NKikimr
