+++ ydb/core/driver_lib/run/ya.make
@@ -17,8 +17,6 @@ SRCS(
     run.h
     service_initializer.cpp
     service_initializer.h
-    cert_auth_props.h
-    cert_auth_props.cpp
 )
 
 PEERDIR(
@@ -70,7 +68,7 @@ PEERDIR(
     ydb/core/graph/shard
     ydb/core/grpc_services
     ydb/core/grpc_services/base
-    ydb/core/grpc_services/auth_processor
+    # ydb/core/grpc_services/auth_processor
     ydb/core/health_check
     ydb/core/http_proxy
     ydb/core/jaeger_tracing
