--- ydb/core/grpc_services/auth_processor/ya.make
@@ -1,11 +1,0 @@
-LIBRARY()
-
-SRCS(
-    dynamic_node_auth_processor.cpp
-)
-
-PEERDIR(
-    contrib/libs/openssl
-)
-
-END()
