+++ ydb/core/grpc_services/ya.make
@@ -110,7 +110,7 @@ PEERDIR(
     ydb/core/grpc_services/counters
     ydb/core/grpc_services/local_rpc
     ydb/core/grpc_services/cancelation
-    ydb/core/grpc_services/auth_processor
+    # ydb/core/grpc_services/auth_processor
     ydb/core/health_check
     ydb/core/io_formats/ydb_dump
     ydb/core/kesus/tablet
