+++ ydb/core/security/certificate_check/ut/ya.make
@@ -1,0 +1,19 @@
+UNITTEST_FOR(ydb/core/security/certificate_check)
+
+FORK_SUBTESTS()
+
+TIMEOUT(600)
+
+SIZE(MEDIUM)
+
+PEERDIR(
+    ydb/core/testlib/default
+)
+
+YQL_LAST_ABI_VERSION()
+
+SRCS(
+    cert_utils_ut.cpp
+)
+
+END()
