+++ ydb/services/ydb/ya.make
@@ -22,7 +22,7 @@ PEERDIR(
     ydb/core/formats
     ydb/core/grpc_services
     ydb/core/grpc_services/base
-    ydb/core/grpc_services/auth_processor
+    ydb/core/security
     ydb/core/grpc_streaming
     ydb/core/protos
     ydb/core/scheme
